USB: fix error when EP0 packet size is multiple of EP0_MAX_PACKET_SIZE
Max packet size should be 32 to match the hardware FIFO length
When a transaction finishes the final packet should have less
than EP0_MAX_PACKET_SIZE bytes. If the final packet is exactly
EP0_MAX_PACKET_SIZE bytes long, as is the case with transactions
of multiples of EP0_MAX_PACKET_SIZE bytes, then an empty packet is
sent at the end to close the transaction.