Add Microsoft WCID descriptor
Add the MS string descriptor at idx 0xEE, which in turn
points to a vendor request of 0xFD. This vendor request
returns a buffer that notifies Windows 7/8/10 that we
use the WINUSB driver. This stops the user from having
to set it up with Zadig, and the device shoudl work out
of the box on those platforms.

There is useful documentation on this here:
https://github.com/pbatard/libwdi/wiki/WCID-Devices

Additionally I have changed bInterfaceClass, bInterfaceSubClass,
and bInterfaceProcotol to 0xFF (user defined). Their previous
values were marked as "reserved for future standardization." in
the USB spec.