Py3 simple (#78)
next phase of Python3 support, including lots of unittests and a FakeRfCat simulated RfCat dongle for unittesting the Python APIs.
many thanks to @PlantDaddy for getting us this far already!  this is a light lift.  some of the changes (mostly Vstructs) are back-ported from Kali's current, but broken rendition of RfCat.

* minor tweaks which seem to make py3 pretty happy.  haven't changed the ipython line over to py3, but that's next.

* vstructs fixes and final tweaks for py3 in rfcat and chipcon_*py

* more py3 fixes, fakedongle, and unittest work

* bugfixes.  still working on unittests

* py3 for fakedongle

* intelhex py3 tweaks

* bugfix: ep0 string/bytes

* unittest work

* more work on fakedongle and unittests.  removed test_fakedongle since fakedongle needs to be it's own thing.

* unit tests and py3 bugfixes

* more bugfixes and unittests!

* bugfixes (py3 *and otherwise*)
rearranging NIC_XMIT_LONG to NIC_LONG_XMIT on python to correspond to the firmware defines.
improved getters to return real values, not the packet received from the dongle.

* lots of API unittests, and lots of fakeDongle enhancements to support them.