Fix calculatePktChanBW
The comments in this function state this:
''' calculates the optimal ChanBW setting for the current freq/baud
        * totally experimental *
        from Smart RF Studio:
        1.2 kbaud BW: 63khz
        2.4 kbaud BW: 63khz
        38.4kbaud BW: 94khz
        250 kbaud BW: 600khz
        '''
The 1.2kbaud listed 63khz, yet the table of possibles includes 53e3. Not that big a difference, but when passing 53e3 to setMdmChanBW, it always triggers an exception in that function. 

Removing 53e3 from the table is in line with the comments and prevents triggering the exception in setMdmChanBW